jwt installed and working

+++ part4/bloglist/controllers/blogs.js
@@ -4,14 +4,6 @@ const blogRoutes = require("express").Router();
 const Blog = require("../models/blog");
 const User = require("../models/user");
 
-const getTokenFrom = (request) => {
-  const authorization = request.get("authorization");
-  if (authorization && authorization.startsWith("Bearer ")) {
-    return authorization.replace("Bearer ", "");
-  }
-  return null;
-};
-
 blogRoutes.get("/", async (request, response) => {
   const blogs = await Blog.find({}).populate("user", { username:1,user:1 });
   return response.json(blogs);
@@ -21,12 +13,12 @@ blogRoutes.post("/", async (request, response) => {
   const { title, author, url } = request.body;
 
   // eslint-disable-next-line no-undef
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET);
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
   if (!decodedToken.id) {
     return response.status(401).json({ error: "token invalid" });
   }
-  const user = await User.findById(decodedToken.id);
 
+  const user = await User.findById(decodedToken.id);
   const blog = new Blog({
     title, author, url, user: user._id
   });
@@ -38,6 +30,13 @@ blogRoutes.post("/", async (request, response) => {
 });
 
 blogRoutes.delete("/:id", async (request, response) => {
+
+  // eslint-disable-next-line no-undef
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
   const id = request.params.id;
   await Blog.findByIdAndDelete(id);
   return response.status(204).end();
@@ -45,6 +44,13 @@ blogRoutes.delete("/:id", async (request, response) => {
 
 
 blogRoutes.put("/", async (request, response) => {
+
+  // eslint-disable-next-line no-undef
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  }
+
   const newBlog = request.body;
   const { title, author, url, likes, id } = newBlog;
